Refactoring Code and adding Documentation
