+++ documentation/mkdocs.yml
@@ -77,6 +77,7 @@ nav:
         - containerExecuteStructureTests: steps/containerExecuteStructureTests.md
         - containerPushToRegistry: steps/containerPushToRegistry.md
         - debugReportArchive: steps/debugReportArchive.md
+        - deployIntegrationArtifact: steps/deployIntegrationArtifact.md
         - detectExecuteScan: steps/detectExecuteScan.md
         - dockerExecute: steps/dockerExecute.md
         - dockerExecuteOnKubernetes: steps/dockerExecuteOnKubernetes.md
